ref(ui) Consolidate onboarding state styles
Use the same components for the releases & performance onboarding states.
This new component may see more action in the future as onboarding
states are revisited.
