+++ src/sentry/static/sentry/app/views/releases/list/releaseLandingCard.jsx
@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import styled from '@emotion/styled';
 
+import OnboardingPanel from 'app/components/onboardingPanel';
 import space from 'app/styles/space';
 import {t} from 'app/locale';
 import Button from 'app/components/button';
@@ -29,50 +30,26 @@ class ReleaseLandingCard extends React.Component {
     const {card, cardsLength, step} = this.props;
     const finalStep = step === cardsLength - 1;
     return (
-      <Container>
-        <IllustrationContainer>
-          <CardComponentContainer>{card.svg}</CardComponentContainer>
-        </IllustrationContainer>
-
-        <StyledBox>
-          <h3>{card.title}</h3>
-          {card.disclaimer && <Disclaimer>{card.disclaimer}</Disclaimer>}
-          <p>{card.message}</p>
-          {finalStep ? (
-            <Button
-              href="https://docs.sentry.io/learn/releases/"
-              onClick={this.props.onClick}
-            >
-              {this.getMessage()}
-            </Button>
-          ) : (
-            <Button onClick={this.props.onClick}>{this.getMessage()}</Button>
-          )}
-        </StyledBox>
-      </Container>
+      <OnboardingPanel
+        image={<CardComponentContainer>{card.svg}</CardComponentContainer>}
+      >
+        <h3>{card.title}</h3>
+        {card.disclaimer && <Disclaimer>{card.disclaimer}</Disclaimer>}
+        <p>{card.message}</p>
+        {finalStep ? (
+          <Button
+            href="https://docs.sentry.io/learn/releases/"
+            onClick={this.props.onClick}
+          >
+            {this.getMessage()}
+          </Button>
+        ) : (
+          <Button onClick={this.props.onClick}>{this.getMessage()}</Button>
+        )}
+      </OnboardingPanel>
     );
   }
 }
-
-const Container = styled('div')`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-  min-height: 450px;
-  padding: ${space(1)};
-`;
-
-const StyledBox = styled('div')`
-  flex: 1;
-  padding: ${space(3)};
-`;
-
-const IllustrationContainer = styled(StyledBox)`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
 
 const CardComponentContainer = styled('div')`
   width: 550px;
